// packages/discord/src/operations/listPins.ts
export const ListPinsInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  channel_id: Schema.String.pipe(T.PathParam()),
  before: Schema.optional(Schema.String),
  limit: Schema.optional(Schema.Number),
}).pipe(
  T.Http({ method: "GET", path: "/channels/{channel_id}/messages/pins" }),
);
export type ListPinsInput = typeof ListPinsInput.Type;
// ...
export type ListPinsOutput = typeof ListPinsOutput.Type;

// The operation
export const listPins = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: ListPinsInput,
  outputSchema: ListPinsOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));